Committed work in progress including "Symbiote" sketch.
