+++ package.json
@@ -7,6 +7,9 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
+    "create-react-class": "^15.6.3",
+    "exports-loader": "^1.1.0",
+    "infusion": "^3.0.0-dev.20200326T173810Z.24ddb2718",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-scripts": "3.4.1"
