+++ public/index.html
@@ -7,6 +7,7 @@
     <meta name="theme-color" content="#000000" />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+
     <title>React Launchpad Slider Puzzle</title>
   </head>
   <body>
